freebsd-relayd: configure: Add -pthread to LDFLAGS
One of the reasons for the -pthread flag in LDFLAGS is to fix linking of
relayd against a static version of libcrypto. When relayd is linked
against a LibreSSL libcrypto.a, it crashes shortly after starting. The
crashing subprocess is hce and the error message is:

    hce: hce_setup_events: tls_config_new

A tested workaround is to link against libcrypto.so, but that is not an
option if LibreSSL is not the default SSL implementation on the system.