Explicitly flush redirect rules after reloading
If a config update disabled a redirect, the corresponding NAT rule was
not getting flushed.  This is because the F_ACTIVE_RULESET flag does not
get preserved across reloads, so pfe_sync() doesn't know to disable the
rule.

Update init_tables() to handle this.  All of this state management seems
overly complicated and due for a rewrite, but let's fix the existing
code first.